refactor : move method to 'Subwaypath' and remove unused class 'TimePath'
