+++ src/main/java/nextstep/subway/maps/map/domain/TimePaths.java
@@ -1,25 +1,24 @@
 package nextstep.subway.maps.map.domain;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class TimePaths {
 
-    private final List<TimePath> paths;
+    private final List<SubwayPath> paths;
 
     private TimePaths(List<SubwayPath> paths) {
-        this.paths = paths.stream()
-                .map(TimePath::new)
-                .collect(Collectors.toList());
+        this.paths = Collections.unmodifiableList(new ArrayList<>(paths));
     }
 
     public static TimePaths of(List<SubwayPath> paths) {
         return new TimePaths(paths);
     }
 
-    public TimePath findFastestArrivalPath(LocalDateTime departTime) {
+    public SubwayPath findFastestArrivalPath(LocalDateTime departTime) {
         return this.paths.stream()
                 .min(Comparator.comparing(timePath -> timePath.getArrivalTime(departTime)))
                 .orElseThrow(() -> new RuntimeException("경로가 존재하지 않는 요청입니다."));
